I ran ./configure CPPFLAGS='-Wall --pedantic --W' to generate a bunch more warnings. Cleaned up most, BUT some of the sign compare warnings in the db* code I didn't touch because I was worried I'd break things.  I didn't want to cast if it wasn't clear it was OK.
There are not many left, so it would be helpful to have someone else look
at the output from make when run with the above CPPFLAGS.

I didn't bother "fixing" the C++ style comments.  What compiler can't deal
with those.  I do think the casts and mixed sign compares might get us in
trouble some day if any numbers got very large (e.g. a seek position).

+++ src/error.c
@@ -271,7 +271,7 @@ char *getErrorString(int number)
     int i;
     static char message[50];
     
-    for (i = 0; i < sizeof(swishErrors) / sizeof(swishErrors[0]); i++)
+    for (i = 0; i < (int)(sizeof(swishErrors) / sizeof(swishErrors[0])); i++)
         if ( number == swishErrors[i].error_num )
             return swishErrors[i].message_string;
 
@@ -302,7 +302,7 @@ int SwishCriticalError(SWISH *sw)
     if ( !sw )
         return 1;
     
-    for (i = 0; i < sizeof(swishErrors) / sizeof(swishErrors[0]); i++)
+    for (i = 0; i < (int)(sizeof(swishErrors) / sizeof(swishErrors[0])); i++)
         if ( sw->lasterror == swishErrors[i].error_num )
             return swishErrors[i].critical;
